Version 1.19+5
Jerk is no longer applied to the boundaries between travel moves and
printing moves
On the Duet WiFi, the M122 report includes the network reconnect count
The progress bar on PanelDue and the estimated end time based on
filament usage were getting slightly ahead of the true values,
especially for small gcode files
Fixed a bug whereby if a move had a tiny amount of XY movement and a
much larger amount of extrusion, an extrusion speed in excess of the
extruder speed limit set in M203 might be used. This could also be
triggered if a move that was too short to need any steps was followed by
a pure extrusion move.