Version 1.21RC2
The filament usage comments generated by Ideamaker are now recognised
A separate Z probe type (P9 in M558) is now used for BLTouch and
compatible Z probes. When this is selected, the deployprobe.g macro is
run just before each probing move, and retractprobe.g is run every time
the probe triggers, or at the end of the probing move if it doesn't
trigger
All heaters are turned off during probing moves if parameter B1 is
specified in the M558 command
The M671 command can be used with just one pair of X and Y coordinates,
which disables the manual bed calibration assistant
Continuous rotation proximal and distal joints are now supported on
SCARA printers, and continuous rotation turntables are assumed on Polar
printers
When beginning a print from SD card, macro /sys/start.g is run if it is
present
Sending M0 or M1 from a source other than the SD card no longer stops an
SD card print that is not paused
When a serial input channel detects a framing or overrun error other
than in a GCode end-of-line comment, the line of GCode in which it
occurs is discarded
When the M207 Z hop setting was changed during a print while a travel
move is in progress, at the end of the travel move the head was lowered
by the new value for Z hop instead of the original value
Endstop pins 5-9 (E2-E6) on the expansion connector are now working
If neither of the /www/reprap.htm and /www/html404.htm files was present
when the web server as asked to fetch a file, it returned a 404 error,
however there was no associated plain text response so most browsers
displayed a blank screen (fixed in Duet WiFi/Ethernet build only)
If a G0 or G1 command resulted in no movement (e.g. because it was a
duplicate of a previous line) then correct synchronisation between
subsequent movement and non-movement commands was lost