On the way to 2.02RC7
HTML responder changes for new DWC
Failure to delete a file no longer generates a global error message if the reason is that the file or path was not found. This was done for the Cura upload plugin, and also means that the 'silent' parameter to the delete call is no longer needed.