Add array of active and standby temps to status reports
In addition to current temnperatures, the temps object now
contains an array of active temps and an array of standby
temps indexed by global heater number.

Since standby temp was just added to the bed object the standby
temps were also added to both the chamber and cabinet objects for
completeness.