Makes origin ref point of motor pos and buildup
Having the origin as motor position point of reference makes
the buildup compensation equations simpler.

Spool buildup reference point is set to the origin as well.
So spoolRadii-values (M669 R) is assumed to include the
buildup that is on spools when nozzle at origin.

The common reference point makes the buildup
compensation equations simpler yet.

Behaviour changes:
 - M114 motor pos print outs gets easier to interpret
 - Users should no longer configure mountedLine (M669 W)