Skip file I/O on heightmap ops if UsingLinuxInterface() is true
The SaveHeightMap and TrySaveHeightMap functions were trying to
operate on a non-existent SD card when being controlled by the DSF.
Because the commands failed, the DSF wouldn't save to the virtual
SD card either.  Those functions (and LoadHeightMap for completeness),
now check UsingLinuxInterface() and skip the file I/O if true.