Re-implementation of ST7567 support for Duet 2 Maestro on v3.02 (#424)
* Adapted the latest version (2020-08-07) of v3.02-dev to include support
for a ST7565 minipanel display in the Duet 2 Maestro. An experimental
contrast parameter "C" (range 0-100) has been added to the M918 as well.
This builds on previous code where a ST7576 display was made to work on
a Duet 3 Mini board.

* Added final comments, TODO's and clarifications

* Added a default define to set the LCD_CS on the Maestro high permanently
on display init. LCD_CS functions as a gate pin for ST7567 displays on
the Maestro, and that doesn't have to happen with every chip select. In
this specific case, EXP_1 functions as the alternative LCD_CS, since the
original LCD_CS cannot provide an active low.