Allow zero values again in M913 (also M906 and M917)
Prior to version 3.5beta1 there were no checks on zero or nbegatove
parameters in M906, M913 and M917. In 3.5beta1 a check was added that
these values are positive. However, we give an example power fail script
that uses M913 to sert zero current. So not we allow zero again, ut not
negative values.