Fixed 3rd order motion short segments issue
When allocating moves frmo a 3rd order plan some very short segments
could be created at the start and/or end of a move, if the move boundary
was very close to a phase boundary in the plan. This was inefficient and
resulted in the speed calulation of such segments being inaccurate
because of the rounding of segment times to integral numbers of step
clocks. We now eliminate very short segments by adding their distances
and times to adjacent segments.